pyproject.toml
[project]
name = "incident_manager"
version = "0.1.3"
description = ""
authors = [
    {name = "mathieu", email = "mat@myframboise.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django-quill-editor (>=0.1.42,<0.2.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
django = "5.1.6"
django-environ = "0.11.2"
django-storages = "1.14.2"
gunicorn = "21.2.0"
mysqlclient = "^2.2.7"
openpyxl = "3.1.5"
pillow = "10.2.0"
psycopg2_binary = "^2.9.10"
python = "^3.12"
python-dateutil = "2.8.2"
python-dotenv = "1.0.1"
whitenoise = "6.6.0"

[tool.poetry.group.dev.dependencies]
pylint = "^2.17"
pylint-django = "^2.5"
flake8 = "^6.0"
isort = "^5.12"
djlint = "^1.31"

[tool.isort]
profile = "django"
line_length = 88
multi_line_output = 3

[tool.poetry.scripts]
test = "scripts:tests"
lint = "scripts:lint"
dev  = "scripts:dev"
migrate = "scripts:migrate"
translate = "scripts:translate"
